requirements.txt
flask>=3.0.0
openai>=1.0.0
anthropic>=0.18.0
google-generativeai>=0.3.0
pillow>=10.0.0
werkzeug>=3.0.0